chore: merge createProject with initHardhat
