--- v-next/hardhat/test/internal/cli/init/project-creation.ts
@@ -1,8 +1,0 @@
-import { describe, it } from "node:test";
-
-describe("project-creation", function () {
-  it.todo(
-    "should use the last version of solidity in the template for the hardhat.config.ts file",
-    function () {},
-  );
-});
